service layer enhancement

+++ src/main/java/com/hexaware/casestudy/simplyfly/service/RouteServiceImp.java
@@ -2,10 +2,16 @@ package com.hexaware.casestudy.simplyfly.service;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.hexaware.casestudy.simplyfly.dto.route.RouteAddingRequestDto;
+import com.hexaware.casestudy.simplyfly.dto.route.RouteResponseDto;
 import com.hexaware.casestudy.simplyfly.entity.Route;
 import com.hexaware.casestudy.simplyfly.exception.RouteNotFoundException;
+import com.hexaware.casestudy.simplyfly.exception.ServiceNotAllowedException;
+import com.hexaware.casestudy.simplyfly.mapper.RouteMapper;
+import com.hexaware.casestudy.simplyfly.repository.FlightScheduleRepository;
 import com.hexaware.casestudy.simplyfly.repository.RouteRepository;
 
 import jakarta.transaction.Transactional;
@@ -14,49 +20,68 @@ import jakarta.transaction.Transactional;
 @Transactional
 public class RouteServiceImp implements IRouteService {
 	
+	@Autowired
 	private RouteRepository repository;
+	
+	@Autowired
+	private FlightScheduleRepository flightScheduleRepository;
 
 	@Override
-	public List<Route> getAllRoutes() {
+	public List<RouteResponseDto> getAllRoutes() {
 
-		return repository.findAll();
+		return repository.findAll().stream().map(entity->{
+			return RouteMapper.routeToRouteResponseDto(entity);
+		}).toList();
 	
 	}
 
 	@Override
-	public Route getRouteById(int id) throws RouteNotFoundException{
+	public RouteResponseDto getRouteById(int id) throws RouteNotFoundException{
 		
-		return repository.findById(id).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+		Route route = repository.findById(id).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
 		
-	}
-
-	@Override
-	public Route getRouteBySourceAndDestination(String sourceAirportCode, String destinationAirportCode) throws RouteNotFoundException{
-
-		return repository.findBySourceAirportCodeAndDestinationAirportCode(sourceAirportCode, destinationAirportCode).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+		return RouteMapper.routeToRouteResponseDto(route);
 		
 	}
 
 	@Override
-	public Route addRoute(Route route) {
+	public RouteResponseDto getRouteBySourceAndDestination(String sourceAirportCode, String destinationAirportCode) throws RouteNotFoundException{
 
-		return repository.save(route);
+		Route route = repository.findBySourceAirportCodeAndDestinationAirportCode(sourceAirportCode, destinationAirportCode).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+		
+		return RouteMapper.routeToRouteResponseDto(route);
 		
 	}
 
 	@Override
-	public Route updateRoute(Route route) throws RouteNotFoundException{
-
-		repository.findById(route.getId()).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+	public RouteResponseDto addRoute(RouteAddingRequestDto routeDto) {
 		
-		return repository.save(route);
+		Route route = RouteMapper.routeAddingRequestDtoToRoute(routeDto);
+
+		return RouteMapper.routeToRouteResponseDto(route);
 		
 	}
 
 	@Override
-	public String deleteRouteById(int id) throws RouteNotFoundException{
+	public RouteResponseDto updateRoute(RouteAddingRequestDto routeDto,int id) throws RouteNotFoundException,ServiceNotAllowedException{
 
 		Route route = repository.findById(id).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+		
+		if(flightScheduleRepository.findByRoute_Id(id) != null)throw new ServiceNotAllowedException("updating route with active schedule is not allowed");
+		
+		route = RouteMapper.routeAddingRequestDtoToRoute(routeDto);
+		route.setId(id);
+		
+		return RouteMapper.routeToRouteResponseDto(repository.save(route));
+		
+	}
+
+	@Override
+	public String deleteRouteById(int id) throws RouteNotFoundException,ServiceNotAllowedException{
+
+		Route route = repository.findById(id).orElseThrow(()-> new RouteNotFoundException("Route record not found."));
+		
+		if(flightScheduleRepository.findByRoute_Id(id) != null)throw new ServiceNotAllowedException("deleting route with active schedule is not allowed");
 		
 		repository.delete(route);
